removing added tutorial source code: docs_src/path_params/tutorial001.py, corresponding test update
