--- docs_src/path_params/tutorial003a.py
@@ -1,13 +1,0 @@
-from fastapi import FastAPI
-
-app = FastAPI()
-
-
-@app.get("/users/{user_id}")
-async def read_user(user_id: str):
-    return {"user_id": user_id}
-
-
-@app.get("/users/me")
-async def read_user_me():
-    return {"user_id": "the current user"}
